refactor: organism에서 사용하는 상수의 경로 변경
