+++ app/components/organisms/gallery/GallerySection.tsx
@@ -3,79 +3,10 @@ import GalleryCard from './GalleryCard';
 import { Heading } from '@/components/atoms/text/TextFactory';
 import { MAX_CONTENT_WIDTH } from '@/styles/sizes';
 import { type ReactNode } from 'react';
-
-import image1 from '@assets/images/sample/image-1.png';
-import image2 from '@assets/images/sample/image-2.png';
-import image3 from '@assets/images/sample/image-3.png';
-import image4 from '@assets/images/sample/image-4.png';
-import image5 from '@assets/images/sample/image-5.png';
-import image6 from '@assets/images/sample/image-6.png';
-import image7 from '@assets/images/sample/image-7.png';
-import image8 from '@assets/images/sample/image-8.png';
-import image9 from '@assets/images/sample/image-9.png';
-import image10 from '@assets/images/sample/image-10.png';
-import image11 from '@assets/images/sample/image-11.png';
-
-const cardData = [
-  {
-    title: '2024-1 APPTIVE X WAP<br/>연합 해커톤',
-    url: '',
-    image: image1,
-  },
-  {
-    title: '2024-1 아이디어톤 / 팀빌딩',
-    url: '',
-    image: image2,
-  },
-  {
-    title: '2024-1 MT',
-    url: '',
-    image: image3,
-  },
-  {
-    title: '2024-1 현직자 간담회',
-    url: '',
-    image: image4,
-  },
-  {
-    title: '2024-1 정기 세미나',
-    url: '',
-    image: image5,
-  },
-  {
-    title: '2024-1 협업툴 워크샵',
-    url: '',
-    image: image6,
-  },
-  {
-    title: '2024-1 신입기수 오리엔테이션',
-    url: '',
-    image: image7,
-  },
-  {
-    title: '2024-1 아카데미 학군별 스터디',
-    url: '',
-    image: image8,
-  },
-  {
-    title: '2024-1 부산 연합동아리<br/>네트워킹',
-    url: '',
-    image: image9,
-  },
-  {
-    title: '2024-1 OT&데모데이',
-    url: '',
-    image: image10,
-  },
-  {
-    title: '2023-1 홈커밍데이',
-    url: '',
-    image: image11,
-  },
-];
-
-const evenIndexCards = cardData.filter((_, index) => index % 2 === 0);
-const oddIndexCards = cardData.filter((_, index) => index % 2 !== 0);
+import {
+  cardData,
+  evenIndexCards, oddIndexCards,
+} from '@components/organisms/gallery/galleryCards';
 
 function GallerySection({ id }: { id?: string }) {
   return (
@@ -94,7 +25,7 @@ function GallerySection({ id }: { id?: string }) {
             padding="20px 0"
             cssx={{
               display: 'flex',
-              md: { display: 'none' }, // sm 화면 이상에서는 숨김
+              md: { display: 'none' },
             }}
           >
             {cardData.map((card, index) => (
@@ -104,7 +35,7 @@ function GallerySection({ id }: { id?: string }) {
           <Container
             cssx={{
               display: 'none',
-              md: { display: 'flex' }, // sm 화면 이상에서는 보임
+              md: { display: 'flex' },
             }}
             gap="32px"
           >
